Geometry models with nested geometries are responsible of "cleaning themselves up". Extracted geometry-view-base class to ensure views are removed/cleaned properly when model is removed
